Add `update_fn_prototype` pass using Anvill data
This updates a given subprogram DIE's DW_AT_prototyped attr. and creates a child
DW_TAG_formal_parameter DIE with the specified DW_AT_name for each argument. The
three main shortcomings of this pass are that it doesn't:
  - look for functions by address instead of name
  - create new subprogram DIEs (i.e. doesn't work on stripped binaries)
  - check if a formal_parameter DIE exists for each argument
Also removed `Hints` trait since it doesn't seem like it'll be that useful.