Start adding a location attr to formal parameters
Location attr (i.e. the register/memory used to store a parameter) should be
used to identify parameters since we shouldn't rely on getting names from the
disassembly, but this turned out to not be straightforward since a single
location can be represented in various ways which may involve using call frame
information (i.e. the .eh_frame or .debug_frame sections). Currently
dwarf-writer emits a single DW_OP_regx or DW_OP_bregx as a canonical way of
representing the location. Ghidra accepts this as valid dwarf data but I still
have to test if it uses the location data or gets it from its own analysis. The
remaining issue with this feature is that if a binary has a param DIE in a
non-canonical representation, dwarf-writer will create a duplicate DIE the first
time it processes the binary.