Refactored `ELF` to provide access to DWARF data in a more sensible way
I don't really like how `ELF` owns two copies of the data which are not kept in
sync, but it seems like a more sensible way to get access to the data provided
by object::File and gimli::write::Dwarf than the closure approach since the fn
signatures of passes that operate on DWARF units are more decoupled from the
fields of `ELF`. This quirk is mostly a limitation of object::File which must be
created from a reference, but it's probably acceptable.