Add `DwarfType` which maps onto the way DWARF encodes type info
The idea is that `anvill::Type` is split into variants in a way that makes
parsing the input JSON easy while `DwarfType` mirrors how DWARF encodes type
info. This also makes the methods in `dwarf_unit` and `dwarf_entry` mostly
independent of how the anvill spec formats types. The only method in these two
modules that should be anvill-specific is the top-level `process_anvill`.