Finish handling pointers in EntryRef::init_type
This commit finishes adding support for pointers to pointers or primitive types.
I still need to test this before merging into master, but the anvill data in
tests doesn't exactly match the source (as expected) so it seems like it'll be
easier to manually write anvill JSONs specs and write them to a dummy ELF. On a
related note, it'd be nice to add code coverage profiling for tests to get a
sense of everything being tested.